% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{ppareto1}
\alias{ppareto1}
\title{Cumulative distribution function of the Pareto 1 distribution}
\usage{
ppareto1(x, mu, alpha)
}
\arguments{
\item{x}{a (positive) vector}

\item{mu}{a number (the lower bound)}

\item{alpha}{a number (the tail index)}
}
\value{
the c.d.f. of the Pareto 1 distribution at points \code{x}
}
\description{
Cumulative distribution function of the Pareto 1 distribution
}
\examples{
ppareto1(2, 1, 1.5)
}
