% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Table_Top_Share}
\alias{Table_Top_Share}
\title{Table of top shares (using three thresholds)}
\usage{
Table_Top_Share(data, p = 0.01, q1 = 0.1, q2 = 0.05, q3 = 0.01)
}
\arguments{
\item{data}{dataframe obtained from \code{tidy_income} function}

\item{q1}{numeric, the probability level to model a Pareto distribution (default 10% - top 10%)}

\item{q2}{numeric, the probability level to model a Pareto distribution (default 5% - top 5%)}

\item{q3}{numeric, the probability level to model a Pareto distribution (default 1% - top 1%)}
}
\description{
Table of top shares (using three thresholds)
}
