% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.R
\name{tidy_income}
\alias{tidy_income}
\title{Convert income/wealth Data}
\usage{
tidy_income(income, weights)
}
\arguments{
\item{income}{a vector of data (income or wealth)}

\item{weight}{a vector of weight (same length as \code{income}}
}
\value{
a dataframe with 4 columns, \code{y} the vector income (or wealth), \code{weights} the vector of weights, \code{Fw} the cumulated proportion of people (with weights) and \code{Fx} the cumulated proportion of people (without weights)
}
\description{
Convert income/wealth Data
}
