% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.R
\name{Top_Income}
\alias{Top_Income}
\title{Top Income plot}
\usage{
Top_Income(data, p = 0.01, thr = seq(0.85, 0.999, by = 0.001),
  TSlim = NULL, tail = TRUE)
}
\arguments{
\item{data}{dataframe obtained from \code{tidy_income} function}

\item{p}{probability level (default 1%)}

\item{thr}{numeric vector of probability levels to model a Pareto distribution (from 85% up to 99.9%)}

\item{TSlim}{numeric 2-vector, range of y for the plot (default \code{NULL})}

\item{tail}{logical to plot the tail index (default \cote{TRUE})}
}
\description{
Top Income plot
}
