% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{TopShare}
\alias{TopShare}
\title{Estimate Top Share}
\usage{
TopShare(data, p = 0.1, q = 0.1, method = "edf", epd.direct = TRUE)
}
\arguments{
\item{data}{a vector of observations}

\item{p}{(default \code{0.1})}

\item{q}{(default \code{0.1})}

\item{method}{(default \code{"edf"}, but can be \code{"pareto1"} or \code{"gpd"})}

\item{weight}{a vector of weights}

\item{edp.direct}{logical (default \code{TRUE})}
}
\value{
estimation of the share of income/wealth owned by the top p percent of the population, assuming that the top q percent of the distribution is Pareto distributed
}
\description{
Estimate Top Share
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
TopShare(data_1)
}
