% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{dpareto1}
\alias{dpareto1}
\title{import HMisc
importFrom("graphics", "abline", "legend", "lines", "par", "plot")
importFrom("stats", "integrate", "optim", "optimise", "runif", "var")
importFrom("utils", "install.packages")
Density of the Pareto 1 distribution}
\usage{
dpareto1(x, mu, alpha)
}
\arguments{
\item{x}{a (positive) vector}

\item{mu}{a number (the lower bound)}

\item{alpha}{a number (the tail index)}
}
\value{
the density of the Pareto 1 distribution at points \code{x}
}
\description{
import HMisc
importFrom("graphics", "abline", "legend", "lines", "par", "plot")
importFrom("stats", "integrate", "optim", "optimise", "runif", "var")
importFrom("utils", "install.packages")
Density of the Pareto 1 distribution
}
\examples{
dpareto1(2, 1, 1.5)
}
\seealso{
[ppareto1()], [qpareto1()] and [rpareto1()]
}
