% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Hill}
\alias{Hill}
\title{Hill estimator of the tail index, with weights}
\usage{
Hill(data, weights = rep(1, length(data)))
}
\arguments{
\item{data}{the vector of observations}

\item{weights}{the vector of weights}
}
\value{
Hill estimator of \code{gamma} (inverse of \code{alpha})
}
\description{
Hill estimator of the tail index, with weights
}
\examples{
set.seed(123)
x <- rpareto1(100, 1, 1.5)
w <- rgamma(100,10,10)
Hill(x,w)
}
