% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Table_Top_Share}
\alias{Table_Top_Share}
\title{Table of top shares (using three thresholds)}
\usage{
Table_Top_Share(data, p = 0.01, q1 = 0.1, q2 = 0.05, q3 = 0.01,
  verbose = FALSE)
}
\arguments{
\item{data}{dataframe obtained from \code{tidy_income} function}

\item{p}{probability level (default 0.01)}

\item{q1}{numeric, the probability level to model a Pareto distribution (default 0.1)}

\item{q2}{numeric, the probability level to model a Pareto distribution (default 0.05)}

\item{q3}{numeric, the probability level to model a Pareto distribution (default 0.01)}
}
\description{
Table of top shares (using three thresholds)
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
Table_Top_Share(data_1)$Mat_index
}
