% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Top_Income}
\alias{Top_Income}
\title{Top Income plot}
\usage{
Top_Income(data, p = 0.01, thr = seq(0.85, 0.999, by = 0.001),
  TSlim = NULL, tail = TRUE)
}
\arguments{
\item{data}{dataframe obtained from \code{tidy_income} function}

\item{p}{probability level (default 0.01)}

\item{thr}{numeric vector of probability levels to model a Pareto distribution (from 0.85 up to 0.999)}

\item{TSlim}{numeric 2-vector, range of y for the plot (default \code{NULL})}

\item{tail}{logical to plot the tail index (default \code{TRUE})}
}
\value{
one or two graphs (depending on \code{tail==TRUE})
}
\description{
Top Income plot
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
\dontrun{Top_Income(data_1)}
}
