% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{TopShare}
\alias{TopShare}
\title{Estimate Top Share}
\usage{
TopShare(data, p = 0.1, q = 0.1, method = "edf", epd.direct = TRUE)
}
\arguments{
\item{data}{a vector of observations}

\item{p}{(default \code{0.1})}

\item{q}{(default \code{0.1})}

\item{method}{(default \code{"edf"})}

\item{weight}{a vector of weights}

\item{edp.direct}{logical (default \code{TRUE})}
}
\value{
blah blah
}
\description{
Estimate Top Share
}
